Introduce the MessagingBuilder to build instances of MessageStream
