+++ archive/notifier/src/test/scala/uk/ac/wellcome/platform/archive/notifier/fixtures/NotifierFixture.scala
@@ -2,11 +2,14 @@ package uk.ac.wellcome.platform.archive.notifier.fixtures
 
 import java.net.{URI, URL}
 
+import com.amazonaws.services.sns.model.PublishResult
 import uk.ac.wellcome.messaging.test.fixtures.Messaging
 import uk.ac.wellcome.messaging.test.fixtures.SNS.Topic
 import uk.ac.wellcome.messaging.test.fixtures.SQS.{Queue, QueuePair}
 import uk.ac.wellcome.platform.archive.notifier.Notifier
 import uk.ac.wellcome.platform.archive.common.fixtures.BagIt
+import uk.ac.wellcome.platform.archive.common.messaging.MessageStream
+import uk.ac.wellcome.platform.archive.common.models.NotificationMessage
 import uk.ac.wellcome.platform.archive.common.progress.models.Namespace
 import uk.ac.wellcome.storage.fixtures.S3
 import uk.ac.wellcome.test.fixtures.TestWith
@@ -24,12 +27,17 @@ trait NotifierFixture extends S3 with Messaging with BagIt {
     withActorSystem { actorSystem =>
       withMaterializer(actorSystem) { materializer =>
         withMetricsSender(actorSystem) { metricsSender =>
-          val notifier = new Notifier(
+          val messageStream = new MessageStream[NotificationMessage, PublishResult](
+            actorSystem = actorSystem,
             sqsClient = asyncSqsClient,
             sqsConfig = createSQSConfigWith(queue),
+            metricsSender = metricsSender
+          )
+
+          val notifier = new Notifier(
+            messageStream = messageStream,
             snsClient = snsClient,
             snsConfig = createSNSConfigWith(topic),
-            metricsSender = metricsSender,
             contextUrl = new URL("http://localhost/context.json")
           )(
             actorSystem = actorSystem,
