+++ archive/registrar_async/src/main/scala/uk/ac/wellcome/platform/archive/registrar/async/Main.scala
@@ -21,12 +21,7 @@ object Main extends App with Logging {
   implicit val actorSystem: ActorSystem = AkkaBuilder.buildActorSystem()
   implicit val executionContext: ExecutionContext = actorSystem.dispatcher
 
-  val messageStream = new MessageStream[NotificationMessage, Unit](
-    actorSystem = actorSystem,
-    sqsClient = SQSBuilder.buildSQSAsyncClient(config),
-    sqsConfig = SQSBuilder.buildSQSConfig(config),
-    metricsSender = MetricsBuilder.buildMetricsSender(config)
-  )
+  val messageStream = MessagingBuilder.buildMessageStream[NotificationMessage, Unit](config)
 
   implicit val storageBackend: S3StorageBackend = new S3StorageBackend(
     s3Client = S3Builder.buildS3Client(config)
